working, no holes with threads
