+++ src/generator.py
@@ -13,15 +13,15 @@ segments = 100  # render quality
 # Validation
 assert 70 <= pipe_length <= 250, "pipe_length must be between 70 and 250 mm"
 
-# Derived radius for hex (circumcircle radius)
+# Derived radius for hex (circle radius)
 hex_radius = pipe_diameter / (2 * math.cos(math.pi / 6))
 
 
-def hex_cylinder(h, r, fn=6):
-    return solid2.cylinder(h=h, r=r, fn=fn)
+# def hex_cylinder(h, r, fn=6):
+#     return solid2.cylinder(h=h, r=r, _fn=fn)
 
 
-def simple_thread_profile(diameter, pitch):
+def simple_thread_profile(diameter, _pitch):
     """Create a simplified triangular thread profile (not exact ISO)"""
     return solid2.polygon([[0, 0], [diameter / 2, 0], [diameter / 2, 0.5], [0, 1]])
 
@@ -34,7 +34,7 @@ def simple_thread(diameter, pitch, height):
 
 def hex_pipe():
     # Main hex-shaped body
-    body = solid2.cylinder(h=pipe_length, r=hex_radius, fn=6)
+    body = solid2.cylinder(h=pipe_length, r=hex_radius, _fn=6)
 
     # Threads at both ends
     thread = simple_thread(thread_diameter, thread_pitch, thread_depth)
@@ -49,4 +49,5 @@ def hex_pipe():
 
 if __name__ == "__main__":
     scad = hex_pipe()
-    scad.save_as_scad("hex_pipe.scad", file_header=f"$fn = {segments};")
+    scad.save_as_scad("export/hex_pipe.scad")
+
