+++ tasks/openscad.py
@@ -1,0 +1,9 @@
+from invoke import task
+
+
+@task
+def create(context):
+    print("\n------------")
+    print("Create OpenSCAD File")
+    print("------------\n")
+    context.run("python src/generator.py")
