+++ tasks/tests.py
@@ -1,4 +1,22 @@
+import subprocess
+
 from invoke import task
+
+
+@task
+def openscad_lint(context):
+    print("\n------------")
+    print("OpenSCAD Lint")
+    print("------------\n")
+
+    def check_scad(file_path):
+        try:
+            subprocess.run(["openscad", "-o", "/dev/null", file_path], check=True)
+            print("✅ No syntax errors in SCAD file.")
+        except subprocess.CalledProcessError:
+            print("❌ SCAD file contains syntax errors.")
+
+    context.run(check_scad("export/hex_pipe.scad"))
 
 
 @task
@@ -10,7 +28,7 @@ def pylint(context):
 
 
 @task
-def rufflint(context):
+def ruff_lint(context):
     print("\n------------")
     print("Ruff Lint")
     print("------------\n")
